HIVE-6642 Query fails to vectorize when a non string partition column is part of the query expression (Hari Subramaniyan via Harish Butani)
git-svn-id: https://svn.apache.org/repos/asf/hive/branches/branch-0.13@1583463 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/vector/VectorGroupByOperator.java
@@ -260,7 +260,7 @@ public class VectorGroupByOperator extends GroupByOperator {
             HiveConf.ConfVars.HIVE_VECTORIZATION_GROUPBY_MAXENTRIES);
         this.minReductionHashAggr = HiveConf.getFloatVar(hconf,
             HiveConf.ConfVars.HIVEMAPAGGRHASHMINREDUCTION);
-          this.numRowsCompareHashAggr = HiveConf.getLongVar(hconf,
+          this.numRowsCompareHashAggr = HiveConf.getIntVar(hconf,
             HiveConf.ConfVars.HIVEGROUPBYMAPINTERVAL);
       } 
       else {
